+++ korma/src/commonMain/kotlin/com/soywiz/korma/math/ArrayExt.kt
@@ -1,15 +1,15 @@
 package com.soywiz.korma.math
 
-fun DoubleArray.minOrElse(nil: Double): Double {
+fun FloatArray.minOrElse(nil: Float): Float {
     if (isEmpty()) return nil
-    var out = Double.POSITIVE_INFINITY
+    var out = Float.POSITIVE_INFINITY
     for (i in 0..lastIndex) out = kotlin.math.min(out, this[i])
     return out
 }
 
-fun DoubleArray.maxOrElse(nil: Double): Double {
+fun FloatArray.maxOrElse(nil: Float): Float {
     if (isEmpty()) return nil
-    var out = Double.NEGATIVE_INFINITY
+    var out = Float.NEGATIVE_INFINITY
     for (i in 0..lastIndex) out = kotlin.math.max(out, this[i])
     return out
 }
