+++ korma/src/jvmMain/kotlin/com/soywiz/korma/awt/AwtExt.kt
@@ -19,7 +19,7 @@ fun java.awt.Rectangle.toKorma(out: RectangleInt = RectangleInt()): RectangleInt
     out.also { out.setTo(this.x, this.y, this.width, this.height) }
 
 fun java.awt.geom.Dimension2D.toKorma(out: Size = Size()) =
-    out.also { out.setTo(this.width, this.height) }
+    out.also { out.setTo(this.width.toFloat(), this.height.toFloat()) }
 
 fun Size.toAwt(out: java.awt.geom.Dimension2D = java.awt.Dimension()) =
-    out.also { out.setSize(this.width, this.height) }
+    out.also { out.setSize(this.width.toDouble(), this.height.toDouble()) }
